chore: add aws rum service
