--- src/utils/record-page-view.ts
@@ -1,11 +1,0 @@
-// import { useEffect } from 'react'
-// import { useRouter } from 'next/router'
-
-// function RecordPageView() {
-//   const location = useRouter()
-//   console.log(`location: `, location)
-//   useEffect(() => {
-//     console.log('logging pageview to cwr: ' + location.pathname)
-//     awsRum.recordPageView(location.pathname)
-//   }, [location])
-// }
